Trader: calculate Balance (transaction cost - versement)

+++ app/src/main/java/com/zcdev/pointofsale/data/models/Client.kt
@@ -21,13 +21,14 @@ class Client: Trader {
     }
     constructor()
 
-    fun calculeSV(list_vrs: HashMap<String,Versement>): Int {
-        var sum:Int?=0
+    fun calculeSV(list_vrs: HashMap<String,Versement>): Double {
+        var sum:Double?=0.0
         for ((key, value) in list_vrs) {
             sum = sum!! + value.montant!!
         }
         return sum!!
     }
+
 
 
 }
