+++ app/src/main/java/com/zcdev/pointofsale/data/models/Versement.kt
@@ -5,10 +5,10 @@ open class Versement {
     // generate no-argument constructor by giving default values
     var Id:String?=null
     var date: String?=null
-    var montant: Int?=0
+    var montant: Double?=0.0
 
 
-    constructor(Id:String,date:String, montant:Int){
+    constructor(Id:String,date:String, montant:Double){
         this.Id = Id
         this.date=date
         this.montant = montant
